Document run & backwards compatibility

+++ android/src/com/mygdx/game/Analysis/Data.java
@@ -17,7 +17,7 @@ public interface Data {
      * @return A 2 component integer vector that contains the specified body part's coordinate
      * in indexed frames' screen space.
      */
-    public abstract Vector3 getCoord(int frame, body_part bp);
+    public abstract Vector3 getCoord(long frame, body_part bp);
 
     /**
      * 
@@ -29,7 +29,7 @@ public interface Data {
      * 
      * @return The number of total frames in the videos' data structure.
      */
-    public abstract int getFrameCount();
+    public abstract long getFrameCount();
 
     /**
      *
